delete team redux fixed and working
