+++ src/store/admin/action.js
@@ -7,28 +7,9 @@ import {
   showMessageWithTimeout
 } from '../appState/actions';
 
-export const DELETE_TEAM = 'DELETE_TEAM';
 export const ADD_PLAYERTEAM = 'ADD_PLAYERTEAM'
 
-const deleteTeam = team => ({
-    type: 'DELETE_TEAM',
-    payload: team
-});
-
-export const deleteTeamSuccess = teamId => {
-  return async (dispatch, getState) => {
-    try {
-      const response = await axios.delete(`${apiUrl}/teams/deleteteam/${teamId}`)
-      console.log('TEAM DELETED', response.data)
-      // dispatch(deleteTeam(teamId));
-      dispatch(fetchAllTeams());
-    } catch(error) {
-      console.log(error.message)
-    };
-  }
-};
-
-const addPlayerTeam = (team, player) => ({
+export const addPlayerTeam = (team, player) => ({
   type: 'ADD_PLAYERTEAM',
   payload: team, player
 })
@@ -37,7 +18,6 @@ export const updatePlayerTeam = (teamId, userId) => {
   return async (dispatch) => {
     dispatch(appLoading)
     try {
-      console.log('Dispatch is working')
       const response = await axios.post(`${apiUrl}/teams/${teamId}/add/${userId}`,
       {
         teamId,
